don't try to move partial pixels
On more recent versions of Mac OS X, the mouse cursor position is reported as a
real number rather than an integer. The window position is also specified as a
real number, but gets coerced to an integer (at least on my computer). I think
this has something to do with how high resolution display support was
implemented, and the window position is being rounded to physical pixels.
Anyway, this would cause the cat to be constantly running trying to reach
exactly the mouse cursor. There's likely a better way to do this taking into
consideration the resolution of the display.